workers: move used pocket loigc outside network stats worker
